+++ source/frontend/src/constants/index.ts
@@ -6,6 +6,7 @@ export * from './countryCodes';
 export * from './districtCodes';
 export * from './environment';
 export * from './fileTypes';
+export * from './financialCodeTypes';
 export * from './leaseStatusTypes';
 export * from './noteTypes';
 export * from './organizationIdentifierTypes';
